Adjust the execution order of flush, change usize to u64, rationalize the execution of the tsm part of the statement.

+++ tskv/src/reader/page.rs
@@ -91,7 +91,7 @@ async fn read(
     let page = {
         let _timer = metrics.elapsed_page_scan_time().timer();
         metrics.page_read_count().add(1);
-        metrics.page_read_bytes().add(page_meta.size());
+        metrics.page_read_bytes().add(page_meta.size() as usize);
         reader.read_page(&page_meta).await?
     };
